feat: general fixes

+++ src/model/sensor_input_moving_average_model.h
@@ -9,14 +9,14 @@
 class SensorInputAverageModel : public AddressedVariable {
   private:
   uint8_t _samples;
-  float *_values;
+  int *_values;
   int _index;
-  float _sum;
+  int _sum;
   bool _completed = false;
-  int32_t _maxRangeValue;
-  int32_t _minRangeValue;
+  unsigned int _maxRangeValue;
+  int _minRangeValue;
 
-  std::function<float(uint16_t)> _conversionCallback;
+  std::function<int(uint16_t)> _conversionCallback;
 
   void reset() {
     for (int i = 0; i < _samples; i++) {
@@ -30,13 +30,13 @@ class SensorInputAverageModel : public AddressedVariable {
   public:
   SensorInputAverageModel(u_int16_t address, SemaphoreHandle_t *mutex, uint8_t samples) : AddressedVariable(address, mutex) {
     _samples = samples;
-    _values = new float[_samples];
+    _values = new int[_samples];
     _minRangeValue = (INT32_MIN + 1);
     _maxRangeValue = (INT32_MAX - 1);
     reset();
 
     _conversionCallback = [](uint16_t value) {
-      return static_cast<float>(value);
+      return static_cast<int>(value);
     };
   }
 
@@ -44,18 +44,18 @@ class SensorInputAverageModel : public AddressedVariable {
     delete[] _values;
   }
 
-  void setConversionCallback(std::function<float(uint16_t)> conversionCallback) {
+  void setConversionCallback(std::function<int(uint16_t)> conversionCallback) {
     _conversionCallback = conversionCallback;
   }
 
-  void setValidRange(int32_t min, int32_t max) {
+  void setValidRange(int min, unsigned int max) {
     _minRangeValue = min;
     _maxRangeValue = max;
   }
 
   void addValue(uint16_t newValue) {
     if (xSemaphoreTake(*_mutex, portMAX_DELAY) == pdTRUE) {
-      float parsedValue = _conversionCallback(newValue);
+      int parsedValue = _conversionCallback(newValue);
 
       _sum = _sum - _values[_index] + parsedValue;
       _values[_index] = parsedValue;
@@ -70,7 +70,7 @@ class SensorInputAverageModel : public AddressedVariable {
     }
   }
 
-  float value() const {
+  int value() const {
     return (_sum / _samples);
   }
 
@@ -79,7 +79,7 @@ class SensorInputAverageModel : public AddressedVariable {
   }
 
   bool isOutOfRange() {
-    float currentValue = (_sum / _samples);
+    int currentValue = (_sum / _samples);
     return currentValue <= _minRangeValue || currentValue >= _maxRangeValue;
   }
 };
